Port of the client subdirectory to KF5.
I had to replace the fake_argv and fake_argc variables in lines 539-540 of
kfmclient.cpp because otherwise I got random segfaults (about one time in three)
in the ClientApp constructor. According to QApplication::QApplication()
documentation, it seems that argc must be at least 1 and argv must not be 0.
The same warning existed in the documentation for Qt 4, however, but the KDE 4
version of kfmclient didn't have this problem.

REVIEW: 121853