Fixed a bug when the user presses Reload while viewing a local file or directory
In these occasions, an error message was generated. The reason is that
KonqMainWindow::slotReload used the automatic conversion from a QString
(KonqView::locationBarURL) to QUrl, which generates a QUrl with an empty scheme.

To avoid the issue, QUrl::fromUserInput should be used to convert the QString to
a QUrl, passing QUrl::AssumeLocalFile as third argument, as for remote files
KonqView::locationBarURL() already contains the scheme.

REVIEW: 124429