setQuitOnLastWindowClosed(false) causes lingering processes. Remove.
Seems to be the same reason as in Kate:
Previously, it was set to true again later from
KMainWindowPrivate::init(). I have changed that in KMainWindowPrivate
so that applications have a better chance to set the property as they
like - commit 155f524dd79add7d in kxmlgui.
For Konqueror, true seems to be the correct setting.