Display an error page for error URLs in WebEnginePart
Summary:
When asked to display an error URL, WebEnginePart displayed another
error telling that the web page "error:/..." can't be reached.

To avoid this, add a scheme handler for the "error" scheme to the default
profile. The handler replies with the html code read from an error.html page
with placeholders replaced by the error details contained in the URL.

The code used to produce the error page is taken, with small changes, by
webpage.cpp from KWebKitPart. The error page, too is taken from KWebKitPart and
modified a little.

Test Plan:
Entered an invalid URL in address bar and checked that the error message was
displayed correctly.

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: anthonyfieroni

Differential Revision: https://phabricator.kde.org/D12637