Restore ability to view help pages from Konqueror
Summary:
A generic QWebEngineUrlSchemeHandler which takes data from KIO::storedGet has been added and is now used also for man and info URLs. The old WebEnginePartHtmlMimetypeHandler has been changed into a class which only embeds the content of `file` URLs in the HTML code.

I found out that several calls to `QWebEngineUrlSchemeHandler::requestStarted` may happen before the previous one has replied (it happened for the help protocol where the HTML code produced by KIO in turn contained `help` URLs). Because of this, I needed to implement a queuing mechanism in the new handler.

The new handler can be used as a base class if special processing of data produced by KIO is needed

Test Plan:
- check that man pages displayed with WebEnginePart look as when displayed by KHTML (es: man:/cp)
- check that help pages displayed with WebEnginePart look as when displayed by KHTML (es help:/kate)

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D14233