Fix removing cookies from KCookieServer
Summary:
The call to KCookieServer::deleteCookies always failed because the
cookie name was passed as a QByteArray and not as a QString. This is now
corrected and tests are provided to check the correct behaviour.

The error check for a DBus error in the call to
KCookieServer::deleteCookies also was incorrect, since it called
QDBusAbstractInterface::lastError(), which isn't useful for non-blocking
calls. A correct check using QDBusPendingCall is now used.

Test Plan: run the autotests

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D15753