Fix handling of cookies without a domain
Summary:
QWebEngineCookieStore and KCookieServer handle cookies with empty domain
differently:
- QWebEngineCookieStore gives them a domain which doesn't start with a
dot
- KCookieServer leave them with no domain and stores the host information
in the host field.

Previously, this difference wasn't taken into account. Now, when adding
a cookie from KCookieServer to QWebEngineCookieStore, the cookie's domain
is set according to the host information and, when adding a cookie to
KCookieServer from QWebEngineCookieStore, the domain is removed if not
starting with a dot.

Test Plan: run autotests

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D16139