Fix CMakeLists.txt to compile WebEnginePartHtmlEmbedder on Qt before Qt-5.12
When compiling Konqueror on versions of Qt before 5.12, the
WebEnginePartHtmlEmbedder class is needed to embed local contents in pages
generated by...

Summary:
...WebEnginePartKIOHandler (for example, for the man protocol). The
corresponding source file is added to the the list of source files only if the
Qt version is less than 5.12. However, an extra command in CMakeFiles.txt caused
the file not to be added and the compilation to fail.

Test Plan:
compile Konqueror with Qt older than 5.12.0: the compilation now
completes successfully.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D22736