Fix importing cookies from KCookieServer at startup
Summary:
The cookies imported from KCookieServer were normalized before being
passed to QWebEngineCookieStore::setCookie. Since this function in turn
normalizes cookies, this caused a new cookie with the wrong domain to be
added: the first call to normalize filled the domain field with the
cookie origin host, the second call, seeing a cookie with a non-empty
domain, added a dot at the beginning of the domain.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D23376