Don't install a scheme handler for the "about" protocol
Summary:
The default mimetype for the "about" protocol is "text/html", so
WebEnginePart::attemptInstallKIOSchemeHandler will try to install a
scheme handler for it. This will fail with a warning from Qt because
"about" is an internal protocol.

Test Plan:
configure Konqueror so that it displays the introduction page at
startup, launch it from terminal and check that no warnings regarding the
"about" protocol are displayed.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D23370