Add option to enable or disable internal PDF viewer from Qt 5.13
Summary:
Add an option to allow the user to choose whether or not to display
PDF files opened from web pages using WebEngine.

This doesn't affect PDF files not opened from web pages (for example, files
opened from Dolphin Part).

The default is NOT to use the internal viewer.

The option is at the bottom of the Miscellaneous section of the Web Behavior
page in the configuration dialog. Maybe it should be made more visible.

Test Plan:
go to a web page with a link to a PDF file and click on the link:
with the option disabled Konqueror opens the PDF using a KPart; with the option
enabled, it opens it in the same WebEnginePart.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D25131