Port the About page to QtWebEngine
Summary:
Replace the old About Page based on KHtml with one based on QtWebEngine.

This has been obtained creating a new QWebEngineUrlSchemeHandler for the `konq``
scheme which will display the contents of the about page. The old `about` scheme
can't be used because it's used internally by QtWebEngine.

There are two problems. The first is that if the user's konquerorrc file
contains old default value for the StartURL entry, that value will be used
instead of the new one, which will result in an error message. I don't know how
to force the use of the new default instead of the old one.

The second problem is that clicking on the "Home Folder" link in the launch page
will open the link in the WebEngine part instead of Dolphin part.

Test Plan:
after removing the StartURL entry from the user's konquerorrc file,
launch Konqueror and check that the intro page is displayed. Click on one of the
links in the page (Introduction, Tips, Specifications) and check that the
correct page is displayed.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D26253