Avoid automatically displaying local URLs in WebEnginePart
Summary:
When a file URL is passed to WebEnginePage::acceptNavigationRequest,
don't allow QtWebEngine to decide what to do with it (open it in the
QWebEnginePage or not) but let Konqueror make the choice according to
the user's configuration.

Whitout this change, clicking on the "Home" link in the introduction
page (konq:konqueror) opens the page in WebEnginePart, while it should
open it in DolphinPart (or whichever part the user has chosen to display
directories).

Test Plan:
from the introduction page (konq:konqueror) click the "Home Folder"
link and check that the home folder is opened in DolphinPart and not
WebEnginePart. Create an HTML file with a link to a local image file, open the
HTML file in Konqueror and click on the link to the image. check that the image
is opened in the part chosen to handle images.

Reviewers: dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D26417