Fix crash due to changes in KBookmarkMenu
Summary:
Since version 5.69, KBookmarkMenu doesn't automatically create an action
collection. This causes Konqueror to crash when going on a submenu in
the Bookmarks menu. To avoid it, manually create the action collection.

Also, avoid calling the version of KBookmarkMenu construction which
takes a KActionCollection, as it's deprecated.

Test Plan:
Open the bookmarks menu and hover with the mouse on a submenu. Check
that it crashes. Do the same after this change and check it doesn't crash
anymore.

Reviewers: dfaure, nicolasfella

Reviewed By: dfaure, nicolasfella

Subscribers: arojas, poboiko

Differential Revision: https://phabricator.kde.org/D28789

(cherry picked from commit b61086cfece0d0f11ecfe9c3309697997cada201)