Ensure that WebEnginePart uses proxy settings from KCM
Summary:
Force WebEnginePart to follow the global KDE proxy settings.

QtWebEngine uses the proxy set with QNetworkProxy::setApplicationProxy, which
means that it ignores the KDE proxy settings chosen by the user.

To force QtWebEngine to respect the user's preferences, read the proxy settings
using KProtocolManager and call QNetworkProxy::setApplicationProxy
appropriately.

QtWebEngine only supports a small subset of the options provided by the
proxy KCM: if the current configuration isn't supported, the user is
warned and asked how to proceed.

If the default HTML engine is not QtWebEngine,
QNetworkProxy::setApplicationProxy is still called in case of a
supported configuration, but nothing is done in case of an unsupported
configuration.

Reviewers: dfaure

Differential Revision: https://phabricator.kde.org/D29066