Fix proxy configuration where WebEngine is not the default
Commit 2b45b2af introduced checking and synchronisation of the KIO proxy
configuration with WebEngine (which can only use a limited set of the
available options).  However, if WebEngine is not the default part
(webengineIsDefault in KonqMainWindow::updateProxyForWebEngine()), the
Qt application proxy is unconditionally set to NoProxy.  This may not be
correct if the user has set proxy environment variables suitable for
WebEngine but it is not the default.

In order to cover the case where WebEngine is not the default but the user
may wish to switch to it later, if it is not the default then assume that
the existing proxy settings are as intended and do not change them.