Add support for full screen web pages
Konqueror already has a full screen mode where the menu bar and the
title bar are hidden and the window takes all the space in the screen.
However, this isn't enough as some web pages (for example when those
playing a video) now expect a full screen mode (which I call complete
full screen) where an element of the page, and not just the window,
takes all the screen.

To achieve this kind of full screen, it's necessary to hide not only the
menu bar and the title bar, but also the status bar, the tab bar and the
tool bars. In theory, also the side bar and the terminal emulator should
be hidden, but this commit doesn't to that.

When switching to a different tab, the complete full screen mode is
automatically exited (because it's strictly related to the contents
of the page). Complete full screen mode is NOT entered again
automatically when switching back to the previous page. This is
consistent with the behavior of other browsers such as Chrome or
FireFox.

If Konqueror is exited while in complete full screen mode, its main
window settings are not saved.