Mark the file downloaded from a blob as temporary
When emitting the BrowserExtension::openUrlRequest for the file a blob
URL was downloaded to, add an entry to
KParts::OpenArgumentRequest::metaData. This entry is taken into account
by KonqMainWindow::openUrlRequestHelper which will set the tempFile flag
of the KonqOpenUrlRequest it creates to true.

This way, if the file is opened in Konqueror, it's deleted when the
corresponding tab does. It doesn't have any effect if the file is opened
in an external application.