use ecm_add_test to actuall run sidebar-modulemanagertest
Requiring at least ECM 5.38 comes with a different build dir layout to
help testing of not installed apps. But in combination with the old
add_test syntax ctest can't find the test executable and thus can't run
the test. Using ecm_add_test fixes that. The test is executed with this
change, it still does not pass, though.