Allow the user to choose how to open an URL from history when activating it
Previously, when the user double clicked an entry in the history dialog,
the URL was opened in a new window, with no way to change this
behaviour.

Now the user can configure what to do when activating an entry
choosing from:
- always open the URL in the current tab
- always open the URL in a new tab
- always open the URL in a new window
- open the URL in a new tab unless the current tab has a konq: URL or an
  empty URL, in which cases open the URL in the current tab.

This also changes the behaviour of the history dialog so that the entry
is opened when it is activated, not double clicked.

The user's choice is also honoured from the history sidebar module,
except that the last option can't be used because (as far as I can see),
the module can't access the current URL. In the sidebar module, this
option is considered as if it were "always open in a new tab".