Display different wallet icon depending on whether cached data exists
With the previous improvement to the KWallet integration, the open wallet
icon was displayed in the status bar whenever the current page
contained web forms, regardless of whether or not there was cached data.
This is confusing for the user.

Now if the web page contains forms but there's no cached data for it,
the closed walled icon is displayed instead. This way, the user
knows that there's to possibility to use the KWallet integration for the
current page but that there's no cached data for it.