Add an entry to the wallet popup menu to start filling of web forms
Sometimes the user may need to fill the web forms on a page after the
page has been loaded (for example because the page has unconvetional
ways of asking for login information which prevent automatic filling
from working).

A second entry to show the wallet manager is also added to the menu.