Fix a bug causing being unable to overwrite existing cached data
This bug has been introduced by changing
WebEngineWallet::WebEngineWalletPrivate::saveDataToCache so that it
doesn't delete entries in pendingSaveRequests anymore but leaves this
task to its caller. However, the caller can't know whether the data has
actually been saved or whether the saveFormDataRequested signal has been
emitted because the user needs to decide what to do. As a consequence,
the entry in pendingSaveRequests will always be removed, so that when
the user decides to save it, it won't be there anymore.

To avoid this issue, saveDataToCache now returns a bool telling whether
the data has been saved (and the entry must be remvoed) or not.