Enable option to show hidden folders in sidebar, plus small sidebar code fixes (eg: scrollbar display)
Adds option to use the "ShowHiddenFolders" option in .desktop files for sidebar
buttons, as well as a context menu entry to toggle it when the option is present
in the .desktop file.

Included is a small fix for the width of the column display, which now allows
for the scrollbar to be displayed properly when the tree does not fit the
horizontal area. This is still not perfect though.

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D29871