Improve feature permission request
QtWebEngine only seems to support site-wide rather than page-wide
feature permissions, since the URL passed by the
featurePermissionRequest signal always has / as path. Since the
permission bar was only shown if this URL is equal to the page's URL
(which includes the path), they were almost never shown and the feature
request was automatically denied (except for notifications). This is now
fixed by comparing the request URL with an URL with / as path.

Besides, the feature permission bar always had the same label (referring
to geolocalization). Now the label has a different text according to the
feature.

Multiple permission bars can now be displayed at once.