Fix focus when opening page in a new tab
When opening a page in a new tab by middle-clicking a link, focus wasn't
moved to the new page but remained in the location bar. This happens
because the createNewWindow signal is emitted by
NewWindowPage::acceptNavigationRequest using an empty URL rather than
the true URL. To fix this, the real URL is passed to the createNewWindow signal.