Revert "Fix focus when opening page in a new tab"
This reverts commit a69162a9969abf64125808ff53383fb948c64452.

It seems that this causes issues with links to non-html files. For
example:
- open a web page
- from that page, middle click on a link to another web page to open it
  in another tab
- from the second page, middle click on a link to a PDF file

What should happen is that the PDF file is displayed in a third tab
while the first two tabs remain unchanged. Instead, with original commit
the PDF file is displayed not only in the third tab but also in the
first one.