CMake: handle hunspell-has-no-dictionaries
If there are **no** dictionaries for hunspell, then
the AVAILABLE DICTIONARIES line is followed by no
output (no lines listing dictionaries). Then
the match fails, so CMAKE_MATCH_1 is empty,
and substituting it in the REGEX REPLACE command
triggers a CMake error like this:

CMake Error at cmake/modules/FindHunspell.cmake:54 (string):
  string sub-command REGEX, mode REPLACE needs at least 6 arguments total to
  command.

Quote it, so that at least the empty string is passed
in when no dictionaries are available.