Don't ask to store credentials if all fields are empty and use KActionCollection::setDefaultShortcut
If the user clicked the `submit` button in a form without having filled
any field in the form, if there was no cached data for that URL, he
would be asked whether to save credentials, which makes no sense since
he didn't enter any. This could also happen if a web page offers
alternative ways to login (for instance QR codes): using the alternative
method submits the form while leaving the fields empty, which triggered
the "save credentials" question.

To fix this issue, now we ignore forms which only have empty fields
(read-only fields aren't taken into account). This only happens for URLs
with no cached data associated with it.