Make Konqueror compatible with KIO 5.86.0
Due to a change in KIO::DesktopExecParser::hasSchemeHandler (commit
5fa55a2395cbfb6504e56bf71c869c8e49902e13Q) URLs entered in the
navigation bar either were always opened in a new tab or they were
opened in another program altogether (for example, man and info URLs).

To fix this issue, it has been necessary to copy the implementation of
KRun::init and KParts::BrowserRun::init in KonqRun::init, making the
necessary changes to restore the old behaviour.

BUG: 442636