Port konqueror plugins to use embedded json metadata
Now the metadata is embedded as json in the plugin and we use
KPluginMetaData::findPlugins to locate those plugins.

With the metadata now being included in the binary plugin, those plugins need
to be installed in all the namespaces. This would increase the file duplication.
Instead a symlink to the plugin is installed in the namespaces.
The rc files are installed in a common directory to avoid the need for file
duplication or symlinks.

Consequently we no longer need the desktop to json conversion at runtimme
using KPluginMetaData::fromDesktopFile. This is planned to be deprecated,
see https://phabricator.kde.org/T14778.

To avoid having defunct ode party lying around, the unused methods of the imported KParts::Plugin
class were cleaned up.

This is also a requirement for https://phabricator.kde.org/T14564, because
the conversion should include all files in one go to reduce issues for the translators.

This is a follow-up of https://phabricator.kde.org/T13806.