Avoid having more than one certificate error dialog at the same time in the same window
Because QDialog::open is asynchronous, if a page generated several 
certificate errors, they could be shown at the same time, which could
make it difficult for the user to interact with them.

To avoid this, now dialogs belonging to the same window are queued so
that there can't be more than one of them visible at the same time.