Clean up defunct opensearch classes
We only ever used google, but here the API endpoint was removed.
In the code we silently failed without any indication what went wrong.
See also BUG: 299846.

If we were to add suggestions, we would need to reimplement it anyways.

This way we also get rid of the QtScript dependency.