Make Konqueror have correct icon on Wayland
It seems that on Wayland the icon for Konqueror can't be found because
the org.kde.konqueror.desktop is installed in kservices5 instead of in
applications

In `CMakeLists.txt`, it mentions that there's a reason for installing it there; because of this I installed the file in both places. I know it's not optimal, but I don't know if there's another way to fix the problem on Wayland.

BUG: 452149